Removed intermediate StarfileStack class
Simplified hierarchy - RelionStarfileStack now directly subclasses ImageSource (encountering Starfiles that are not Relion Starfiles deemed very unlikely).
